+++ CHANGES.md
@@ -53,6 +53,7 @@
 * Add MagnetDL torrent provider
 * Add SceneHD torrent provider
 * Add Skytorrents torrent provider
+* Add WorldOfP2P torrent provider
 * Change do not have shows checked by default on import page. To re-enable import shows checked by default,
   1) On config page 'Save' 2) Stop SG 3) Find 'import_default_checked_shows' in config.ini and set '1' 4) Start SG
 * Add Nyaa (.si) torrent provider